# src/main/java/com/example/bankcards/dto/cards/TransferResponse.java
package com.example.bankcards.dto.cards;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Schema(description = "Ответ на запрос перевода средств")
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class TransferResponse {
    @Schema(description = "Данные карты отправителя")
    private CardDTO fromCard;

    @Schema(description = "Данные карты получателя")
    private CardDTO toCard;
}